rename component-files to index.js
