+++ components/ButtonSubmit/index.js
@@ -1,4 +1,4 @@
-export default function ButtonSubmit({ onSubmit }) {
+export default function ButtonSubmit() {
   return (
     <>
       <button type="submit">Submit</button>
