+++ components/PeriotopiaInfo/index.js
